Time to throw this prototype away
Mostly updating the readme and adding some todo notes.

This where I learned that the approach of avoiding implementing as
much CSMS logic as possible is fundamentally flawed. The OCPP protocol
is to complex to be captured in this way without requiring CSMS
implementers to take on far too much protocol-level responsibility.

The new direction I believe this should take is to implement all the
fundamental business logic for managing and monitoring charging
stations and transactions while providing an event-driven API and
a management API for implementing a more sophisticated CSMSs.