Simplify the spec for ocpp_rpc:decode/1
Take advantage of the new types to make this easier to
understand. Also fix the error returned if decoding the message with
jiffy fails. This still is not quite correct, for reasons noted in the
comments, but I don't want to deal with the added complexity of
differentiating between errors decoding the RPC message structure and
errors decoding the request payload at this point.