Launch an ocpp_station for each station in the database.
The idea is that there will always be a gen_statem running that
represents the stations, whether or not they are
connected. Conceptually this makes maintaining state of the station
over reboots and disconnects "simple" because the state machine never
goes away. On the other hand, starting a new state machine for each
connection and stopping it upon disconnect may make the state machines
themselves simpler.

It is probably worth exploring both approaches. The first challenge I
see for the later approach is tracking which PID belongs to which
station's state machine; and doing so in an atomic way. This isn't
such a serious challenge - it can be accomplished by a basic process
regristry that serializes attempts to register new processes. This
could be a fun opportunity to build something using the `via`
gen_statem/gen_server name.