Start a new ocpp_station state machine for every connection
This drastically simplifies the state machine and eliminates a host of
potential race conditions that come from trying to maintain a unique
state machine for each station. We can punt handling duplicate
connections from one station to the CSMS. That is a better place for
it anyway, since the individual station will never have enough
information to decide what it should do and would probably have to ask
the CSMS anyway.