Revert "Start a new ocpp_station state machine for every connection"
While this was a nice idea to simplify the occp_station state machine
and message handling, it was ultimately a mistake. The OCPP
requirements, specifically B04.FR.01 and B04.FR.02 preclude this
because connection loss and resoration are not treated the same as
rebooting a station. Only reboot of the station requires a new
BootNotification Request.

This reverts commit 23eef1572939cb422d4d1370403f3a705f107f92.