Add full supervision tree for station processes
I'm probably getting ahead of myself by including a stub of the
ocpp_evse state machine, but I wanted to see the full supervision tree
spin up.

With this commit almost the full device model (with the exception of
connectors) is modeled as erlang processes. The goal is that wherever
possible the requests from one station can be processed within that
stations supervision tree, limiting concurrent access to the central
CSMS server (that hasn't been designed yet).