Rework the station supervision tree and evse startup
Eliminate the coupling between the ocpp_station_supersup module and
the station database. The thinking is that no stations will be started
when the application starts. The csms server will be responsible for
starting stations dynamically. In addition to completely removing
database access from the single-station related code this makes the
whole application more flexible, since the CSMS can decide which
stations it actually wants to start.

(The websocket handler still needs to be reworked to better decouple
authentication from the database.)