Reimplement ocpp_rpc to improve error handling and message types
This is the first step in reworking the rpc & messages modules to
provide a clearer and more flexible API and datatypes. The
ocpp_requests module will likely be removed in favor of handling all
message operations in one place within, the ocpp_message module. The
separation between the RPC protocol layer in ocpp_rpc module and the
OCPP messages layer is maintained.

Tests are added to check for RPC framework errors where they are
expected.