Remove the websocket server from this application
This should be built as a separate application in order to keep the
core OCPP functionality isolated and allow for different front end
applications or for running the front end in a separate node.

This is only a partial removal - additional components such as the
mnesia database and the JSON validation need to be removed; however,
thier removal (especially the JSON parts) will take more effort.