First cut structure for interaction between a CSMS and ocpp app
Sets up a custom behavior that will be used by implementers of CSMS
applications for interaction between the CSMS and the OCPP
application. Establishes the basic communication structures between
the station fsm and the CSMS client.

This is pretty rough, but it does function in the sense that you can
add stations to the ocpp application.