Refactor: move station mgr & event handler under station supervisor
These process are tightly coupled so it makes more sense to have them
under the same supervision tree. It also means we eliminate the need
to link the ocpp_station fsm to the event manager since if the event
manager (or more likely the station manager) fails the fsm will be
restarted due to the one_for_rest supervision strategy.