Use gproc to register the station, its manager, & its handler
This replaces the ocpp_station_registry and greatly simplifies the
initialization of the station supervision sub-tree.

Still need to consider how ocpp_station_supersup tests if the station
is registered or not. This may actually be something I want to do in
`ocpp_manager` instead of in the ocpp_station_supersup:start_station/3.