Handle errors from handler init/1 funcion by stopping the station
The station shouldn't run if the handler cannot be initialized. By
moving adding the handler in the ocpp_station_manager:init/1 callback,
we can stop the station start process before it finishes and let the
supervisor handle all the cleanup if the handler fails to initialize.

As a bonus, we get a nice informative return value from the
ocpp_station_supersup:start_staion/3 function.