Keep a reference to the message Id when returning sub-objects
This lets us keep a uniform input type (ocpp_message:message/0) for
the ocpp_message API. Otherwise the function would need to be written
to accept "bare" jerk terms or I would have to allow jerk to be used
outside the message module. (I am trying to keep it contained to this
one module to minimize coupling between the ocpp application and its
dependencies. Also in case it is a mistake to use it and I want to
replace it with something else).

Requires a bump in the jerk version we use (rebar.lock).