Respond with an internal error when the ocpp_handler crashes
This requires shifting the some responsibility from the ocpp_handler
module into the ocpp_station_manager in order to deal with a subtle
race condition. The error event must be sent to the state machine from
the station manager _after_ it re-installs the handler. If the error
response is sent before the handler is reinstalled a station that
initiates a new request immediately upon receiving the error response
from the CSMS may never have its request processed. This is because it
is (theoretically) possible for the new request to arrive at the CSMS
and be forwarded to the station event manager before the handler is
re-installed. To avoid this the station manager is responsible for
sending the error event to the state machine.

This creates a bit of extra coupling between the station manager and
the event handler that I'm not really sure I like. I have done my best
to minimize the coupling by having the handler construct the
ocpp_error message prior to exiting, but still it feels like a lot of
coupling. I don't really see a better option though. The only
alternative that immediately comes to mind is to provide a startup
mechanism in the ocpp_handler:init/1 function that would send the
error message, but that doesn't really reduce the coupling at
all. Ultimately these two processes are already pretty strongly
coupled; the manager installs and supervises the handler already.

I could probably take one additional step to reduce the coupling by
adding a function to the handler module API that is called by the
station manager after the handler is re-installed. We could pass the
error reason to that function and then keep all the details about what
the type of the error reason is hidden inside the ocpp_handler module.