Refactor ocpp_handler restart
Move sending the error to the station state machine back into the
ocpp_handler module to reduce coupling between the ocpp_handler module
and the ocpp_station_manager module. Now the manager is just
responsible for supervising the handler, and not responsible for
sending an error to the station state machine (cool, this also recuces
coupling between those two modules as well).