Test disconnect/reconnect while the handler is still processing
This scenario is meand to simulate a badly implemented handler, or
perhaps a long gc pause that causes enough delay for the station to
disconnect and reconnect while the handler is still processing. That
means its response will arrive after a 'new' connection has been made
and is no longer valid.