Re-implement EVSE as a data structure instead of a gen_statem
The EVSE gen_statem was a nice thought, but it doesn't really align
with the requirements or device model of OCPP. The states actually
correspond to the status of each connector attached to the EVSE. By
making this into a data structure instead of a gen_statem the start up
and data for the ocpp_station gen_statem are substantially simplified
as well as producing better alignment with the standard.

The re-implementation is in support of implementing connector status
updates in provisioning process which is included in this commit. This
is the last step for the full "normal" provisioning process (use case
B01, pg. 43, ocpp specification Part 2).

There is still a bit more testing to be done w.r.t. requests to update
the status with invalid EVSE Ids and/or Connector Ids. (Also testing
that the status changes actually happen, but that's not quite so easy).