Make message type visible for use in pattern matching
This is so the type can be used as in function head pattern matching
in the ocpp_station state functions. Using pattern matching will make
the structure of the state machine much more clear and keep from
needing to use case expressions as the outer expression of every
state handler for call events. This aligns better with the OCPP
standard which uses the context (RPCCALL/CALLRESULT) to identify
whether a 'BootNotification' is a request or a response.