Send a SetVariablesRequest to the station
This is the first cut/happy path for this operation. The ability to
send these requests still needs to be added to the `pending` state and
potentially others. What this commit does provide is the basic
skeleton for how requests from the CSMS to the station should work.

This commit reveals a problem with the design of requests in the other
direction. Requests from the station to the CSMS have been implemented
as synchronous calls which is nice for its simplicity; however, this
limits the ability for the connected process to handle a call from the
CSMS while it is waiting for a response from the CSMS. Since the
standard doesn't appear to precluded this we may want to make the
rpccall functions asynchronous, or provide asynchronous versions.