Make ocpp_station:call/2,3 synchronous
Rename the exising implementation of call to 'call_async'. The go-to
should be the synchronous version; however, we still want to provide
an async option since the timeouts for calls to the station may be
very large (the standard suggests using 30 seconds as a default).

This implementation is incomplete - only exists for the idle
state. Still need to add it for the not-idle state and test handling
for errors that occur before the message is sent to the
station (e.g. a call is already pending - should be easy to test this
now that there is a synchronous and an asynchronous version of call/2).