Make ocpp_station:rpccall/2 asynchronous (#8)
By making this operation asynchronous we are able to dramatically
simplify the tests, but also prevent a potential deadlock that could
arise when the station and the CSMS send messages to each other at the
same time. This also means that the websocket layer (when implemented)
can be a simple layer that translates between JSON and `ocpp_message`
terms and forwards all messages to/from the websocket without needing
to concern itself with their content or semantics.