Pre-populate the device model with referenced components & variables
This requires modifying the variable characteristics table to
implement a hierarchy of variables. We must be able to get the
characteristics of a variable defined in a component for all
evse/conductors/instances. As part of this, I also switched from using
binaries as identifiers to strings. The main reason is for
readability, so it might be a mistake.