Use set table types for device model characteristics and attributes
This combines the component identifier and the variable identifier
(and the attribute type for attributes) into a unique key for each
element in the tables. If we used a bag then we would need to add
extra logic to handle changes to the value of an attribute to delete
the previous value. Making these sets with more complex keys removes
this logical complexity in exchange for a little more structural
complexity in the elements of the tables.