Simplify the ocpp_handler behaviour
The previous approach, with one callback per OCPP message type, would
have required up to 21 optional callbacks. This seemed too unwieldy so
I simplified the design to support a handle_ocpp/3 callback. The
downside of this is that it will require implementors to explicitly
NotSupported error for any messages that aren't supported.