Automatically expand the list of EVSE and Connectors
When a StatusUpdate of a report indicates and EVSE ID or a Connector ID that is
not yet known the ocpp_station now automatically expands the list of
evse/connectors to include that evse and any lower ID numbers that are also
unknown. This saves the CSMS from needing to know all details of the stations
prior to provisioning.